Extract snapshot functions into snapshot.go
